requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

mysqlclient==2.2.4

python-dotenv==1.0.1

Pillow==10.4.0

gunicorn==23.0.0
whitenoise==6.7.0